backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
redis
python-dotenv
openai-whisper
transformers
torch
torchvision
torchaudio
deep-translator
pytesseract
Pillow
ollama
psycopg2-binary
chromadb
python-multipart
twilio
